Refactor SelectCustomer and SelectProduct components
- Consolidated SelectCustomer and SelectProduct functionalities into single components.
- Removed SelectCustomers and SelectProducts components to streamline the codebase.
- Updated related imports and usages across the application to reflect the new structure.
- Enhanced SelectCustomer and SelectProduct with improved context management and state handling.
- Added support for single and multiple selection modes in both components.
- Updated GraphQL queries to match the new component names.
- Improved placeholder handling in SelectTagsValue component.
- Adjusted CustomerFields to use the new SelectCustomer component.
