+++ frontend/core-ui/src/modules/types/paths/ContactsPath.ts
@@ -1,8 +1,8 @@
 export enum ContactsPath {
-  Index = 'contacts',
-  Customers = 'customers',
-  Leads = 'leads',
-  Companies = 'companies',
-  Vendors = 'vendors',
+  Index = '/contacts',
+  Customers = '/customers',
+  Leads = '/leads',
+  Companies = '/companies',
+  Vendors = '/vendors',
   Clients = '/clients',
 }
