+++ frontend/libs/ui-modules/src/modules/contacts/components/SelectCustomer.tsx
@@ -1,83 +1,85 @@
-import {
-  AvatarProps,
-  ButtonProps,
-  Combobox,
-  Command,
-  Popover,
-  Skeleton,
-} from 'erxes-ui';
-import { useCustomers } from '../hooks/useCustomers';
-import { useState } from 'react';
+import { SelectCustomerContext } from '../contexts/SelectCustomerContext';
+import { ICustomer } from '../types';
+import { useSelectCustomerContext } from '../hooks/useSelectCustomerContext';
+import { useCustomers } from '../hooks';
 import { useDebounce } from 'use-debounce';
-import { ICustomerInline } from '../types/Customer';
-import React from 'react';
-import { useCustomerInline } from '../hooks';
-import { CustomerInline } from './CustomerInline';
+import { useState } from 'react';
+import { cn, Combobox, Command, Form, Popover } from 'erxes-ui';
+import { CustomersInline } from './CustomersInline';
+import {
+  RecordTablePopover,
+  RecordTableCellContent,
+  RecordTableCellTrigger,
+} from 'erxes-ui';
 
-interface SelectCustomerProps {
-  value?: string;
-  onValueChange?: (value: string) => void;
+interface SelectCustomerProviderProps {
+  children: React.ReactNode;
+  value?: string[] | string;
+  onValueChange?: (value: string[] | string) => void;
+  mode?: 'single' | 'multiple';
 }
 
-export const SelectCustomer = React.forwardRef<
-  React.ElementRef<typeof Combobox.Trigger>,
-  ButtonProps & SelectCustomerProps
->(({ value, onValueChange, children, ...props }, ref) => {
-  const [open, setOpen] = useState(false);
-  const [selectedUser, setSelectedUser] = useState<ICustomerInline | undefined>(
-    undefined,
-  );
+const SelectCustomerProvider = ({
+  children,
+  value,
+  onValueChange,
+  mode = 'single',
+}: SelectCustomerProviderProps) => {
+  const [customers, setCustomers] = useState<ICustomer[]>([]);
+  const customerIds = !value ? [] : Array.isArray(value) ? value : [value];
+  const onSelect = (customer: ICustomer) => {
+    if (!customer) return;
+    if (mode === 'single') {
+      setCustomers([customer]);
+      onValueChange?.(customer._id);
+      return;
+    }
+    const arrayValue = Array.isArray(value) ? value : [];
+    const isCustomerSelected = arrayValue.includes(customer._id);
+    const newSelectedCustomerIds = isCustomerSelected
+      ? arrayValue.filter((id) => id !== customer._id)
+      : [...arrayValue, customer._id];
 
-  const handleSelect = (user: ICustomerInline | undefined) => {
-    setSelectedUser(user);
-    onValueChange?.(user?._id || '');
-    setOpen(false);
+    setCustomers((prevCustomers) => {
+      const customerMap = new Map(prevCustomers.map((c) => [c._id, c]));
+      customerMap.set(customer._id, customer);
+      return newSelectedCustomerIds
+        .map((id) => customerMap.get(id))
+        .filter((c): c is ICustomer => c !== undefined);
+    });
+    onValueChange?.(newSelectedCustomerIds);
   };
-
   return (
-    <Popover open={open} onOpenChange={setOpen} modal>
-      {children ? (
-        children
-      ) : (
-        <SelectCustomerTrigger
-          value={value || ''}
-          selectedUser={selectedUser}
-          setSelectedUser={setSelectedUser}
-          {...props}
-          ref={ref}
-        />
-      )}
-
-      <Combobox.Content>
-        <SelectCustomerList
-          renderItem={(customer) => (
-            <SelectCustomerItem
-              key={customer._id}
-              customer={customer}
-              selectedUser={selectedUser}
-              handleSelect={handleSelect}
-            />
-          )}
-        />
-      </Combobox.Content>
-    </Popover>
+    <SelectCustomerContext.Provider
+      value={{
+        customerIds,
+        onSelect,
+        customers,
+        setCustomers,
+        loading: false,
+        error: null,
+      }}
+    >
+      {children}
+    </SelectCustomerContext.Provider>
   );
-});
+};
 
-export function SelectCustomerList({
-  renderItem,
-}: {
-  renderItem: (customer: ICustomerInline) => React.ReactNode;
-}) {
+const SelectCustomerContent = () => {
   const [search, setSearch] = useState('');
   const [debouncedSearch] = useDebounce(search, 500);
-
-  const { customers, loading, handleFetchMore, totalCount, error } =
-    useCustomers({
-      variables: {
-        searchValue: debouncedSearch,
-      },
-    });
+  const { customerIds, customers } = useSelectCustomerContext();
+  const {
+    customers: customersData,
+    loading,
+    handleFetchMore,
+    totalCount,
+    error,
+  } = useCustomers({
+    variables: {
+      searchValue: debouncedSearch,
+    },
+  });
   return (
     <Command shouldFilter={false}>
       <Command.Input
@@ -85,72 +87,173 @@ export function SelectCustomerList({
         onValueChange={setSearch}
         variant="secondary"
         wrapperClassName="flex-auto"
+        focusOnMount
       />
       <Command.List className="max-h-[300px] overflow-y-auto">
+        {customers?.length > 0 && (
+          <>
+            {customers.map((customer) => (
+              <SelectCustomerCommandItem
+                key={customer._id}
+                customer={customer}
+              />
+            ))}
+            <Command.Separator className="my-1" />
+          </>
+        )}
         <Combobox.Empty loading={loading} error={error} />
-        {customers?.map((customer: ICustomerInline) => renderItem(customer))}
-        <Combobox.FetchMore
-          fetchMore={handleFetchMore}
-          currentLength={customers?.length}
-          totalCount={totalCount}
-        />
+        {!loading &&
+          customersData
+            ?.filter(
+              (customer) =>
+                !customerIds.find((customerId) => customerId === customer._id),
+            )
+            .map((customer) => (
+              <SelectCustomerCommandItem
+                key={customer._id}
+                customer={customer}
+              />
+            ))}
+
+        {!loading && (
+          <Combobox.FetchMore
+            fetchMore={handleFetchMore}
+            currentLength={customersData.length}
+            totalCount={totalCount}
+          />
+        )}
       </Command.List>
     </Command>
   );
-}
+};
 
-export const SelectCustomerTrigger = React.forwardRef<
-  React.ElementRef<typeof Combobox.Trigger>,
-  ButtonProps & {
-    value: string;
-    selectedUser?: ICustomerInline;
-    setSelectedUser: (customer?: ICustomerInline) => void;
-  }
->(({ value, selectedUser, setSelectedUser, ...props }, ref) => {
-  const { loading } = useCustomerInline({
-    variables: { _id: value },
-    skip: !value || selectedUser?._id === value,
-    onCompleted: ({ customerDetail }: { customerDetail: ICustomerInline }) => {
-      setSelectedUser({ ...customerDetail, _id: value });
-    },
-  });
-
-  return (
-    <Combobox.Trigger disabled={loading} {...props} ref={ref}>
-      {value ? (
-        <CustomerInline
-          customer={selectedUser}
-          avatarProps={{
-            size: props.size as AvatarProps['size'],
-          }}
-        />
-      ) : loading ? (
-        <Skeleton className="w-full h-8" />
-      ) : (
-        <Combobox.Value placeholder="Select customer" />
-      )}
-    </Combobox.Trigger>
-  );
-});
-
-export const SelectCustomerItem = ({
-  customer,
-  selectedUser,
-  handleSelect,
-}: {
-  customer: ICustomerInline;
-  selectedUser?: ICustomerInline;
-  handleSelect: (customer?: ICustomerInline) => void;
-}) => {
-  const isSelected = selectedUser?._id === customer._id;
+const SelectCustomerCommandItem = ({ customer }: { customer: ICustomer }) => {
+  const { onSelect, customerIds } = useSelectCustomerContext();
   return (
     <Command.Item
-      key={customer._id}
       value={customer._id}
-      onSelect={() => handleSelect(isSelected ? undefined : customer)}
+      onSelect={() => {
+        onSelect(customer);
+      }}
     >
-      <CustomerInline customer={customer} />
-      <Combobox.Check checked={isSelected} />
+      <CustomersInline customers={[customer]} placeholder="Unnamed customer" />
+      <Combobox.Check checked={customerIds.includes(customer._id)} />
     </Command.Item>
   );
 };
+
+const SelectCustomerInlineCell = ({
+  onValueChange,
+  scope,
+  ...props
+}: Omit<React.ComponentProps<typeof SelectCustomerProvider>, 'children'> & {
+  scope?: string;
+}) => {
+  const [open, setOpen] = useState(false);
+  return (
+    <SelectCustomerProvider
+      onValueChange={(value) => {
+        onValueChange?.(value);
+        setOpen(false);
+      }}
+      {...props}
+    >
+      <RecordTablePopover open={open} onOpenChange={setOpen} scope={scope}>
+        <RecordTableCellTrigger>
+          <SelectCustomer.Value />
+        </RecordTableCellTrigger>
+        <RecordTableCellContent>
+          <SelectCustomer.Content />
+        </RecordTableCellContent>
+      </RecordTablePopover>
+    </SelectCustomerProvider>
+  );
+};
+
+const SelectCustomerRoot = ({
+  onValueChange,
+  className,
+  mode = 'single',
+  ...props
+}: Omit<React.ComponentProps<typeof SelectCustomerProvider>, 'children'> & {
+  className?: string;
+}) => {
+  const [open, setOpen] = useState(false);
+
+  return (
+    <SelectCustomerProvider
+      onValueChange={(value) => {
+        if (mode === 'single') {
+          setOpen(false);
+        }
+        onValueChange?.(value);
+      }}
+      mode={mode}
+      {...props}
+    >
+      <Popover open={open} onOpenChange={setOpen}>
+        <Combobox.Trigger
+          className={cn('w-full inline-flex', className)}
+          variant="outline"
+        >
+          <SelectCustomer.Value />
+        </Combobox.Trigger>
+        <Combobox.Content>
+          <SelectCustomer.Content />
+        </Combobox.Content>
+      </Popover>
+    </SelectCustomerProvider>
+  );
+};
+const SelectCustomerValue = () => {
+  const { customerIds, customers, setCustomers } = useSelectCustomerContext();
+
+  return (
+    <CustomersInline
+      customerIds={customerIds}
+      customers={customers}
+      updateCustomers={setCustomers}
+    />
+  );
+};
+
+const SelectCustomerFormItem = ({
+  onValueChange,
+  className,
+  ...props
+}: Omit<React.ComponentProps<typeof SelectCustomerProvider>, 'children'> & {
+  className?: string;
+}) => {
+  const [open, setOpen] = useState(false);
+
+  return (
+    <SelectCustomerProvider
+      onValueChange={(value) => {
+        onValueChange?.(value);
+        setOpen(false);
+      }}
+      {...props}
+    >
+      <Popover open={open} onOpenChange={setOpen}>
+        <Form.Control>
+          <Combobox.Trigger className={cn('w-full shadow-xs', className)}>
+            <SelectCustomer.Value />
+          </Combobox.Trigger>
+        </Form.Control>
+
+        <Combobox.Content>
+          <SelectCustomer.Content />
+        </Combobox.Content>
+      </Popover>
+    </SelectCustomerProvider>
+  );
+};
+
+export const SelectCustomer = Object.assign(SelectCustomerRoot, {
+  Provider: SelectCustomerProvider,
+  Content: SelectCustomerContent,
+  Item: SelectCustomerCommandItem,
+  InlineCell: SelectCustomerInlineCell,
+  Value: SelectCustomerValue,
+  FormItem: SelectCustomerFormItem,
+});
