+++ frontend/libs/ui-modules/src/modules/contacts/components/index.ts
@@ -1,3 +1,3 @@
-export * from './SelectCustomers';
-export * from './SelectCompanies';
+export * from './SelectCustomer';
+export * from './SelectCompany';
 export * from './CustomersInline';
