+++ frontend/libs/ui-modules/src/modules/products/components/ProductsInline.tsx
@@ -56,7 +56,7 @@ const ProductsInlineEffectComponent = ({
   const { updateProducts, products } = useProductsInlineContext();
   const { products: detailMissingProducts } = useProductsInline({
     variables: {
-      _ids: productIdsWithNoDetails,
+      ids: productIdsWithNoDetails,
     },
   });
