+++ frontend/libs/ui-modules/src/modules/products/components/SelectProduct.tsx
@@ -1,87 +1,358 @@
-import {
-  Combobox,
-  Command,
-  PopoverScoped,
-  usePreviousHotkeyScope,
-} from 'erxes-ui';
+import { SelectProductContext, useSelectProductContext } from '../contexts/SelectProductContext';
+import { IProduct } from '../types/Product';
 import { useProducts } from '../hooks/useProducts';
 import { useDebounce } from 'use-debounce';
-import { useState } from 'react';
-import { IProduct } from 'ui-modules';
-import { ProductInline } from './ProductInline';
-import React from 'react';
+import React, { useState } from 'react';
+import { cn, Combobox, Command, Filter, Form, Popover, PopoverScoped, useFilterContext, useQueryState } from 'erxes-ui';
+import { ProductsInline } from './ProductsInline';
+import {
+  RecordTablePopover,
+  RecordTableCellContent,
+  RecordTableCellTrigger,
+} from 'erxes-ui';
+import { IconShoppingCart } from '@tabler/icons-react';
 
-export const SelectProduct = React.forwardRef<
-  React.ComponentRef<typeof Combobox.Trigger>,
-  React.ComponentPropsWithoutRef<typeof Combobox.Trigger> & {
-    value?: string;
-    onValueChange?: (value: string) => void;
-    scope?: string;
-  }
->(({ value, onValueChange, scope, ...props }, ref) => {
-  const [open, setOpen] = useState(false);
-  const [selectedProduct, setSelectedProduct] = useState<IProduct | undefined>(
-    undefined,
-  );
+interface SelectProductProviderProps {
+  children: React.ReactNode;
+  value?: string[] | string;
+  onValueChange?: (value: string[] | string) => void;
+  mode?: 'single' | 'multiple';
+}
 
-  const { goBackToPreviousHotkeyScope } = usePreviousHotkeyScope();
+const SelectProductProvider = ({
+  children,
+  value,
+  onValueChange,
+  mode = 'single',
+}: SelectProductProviderProps) => {
+  const [products, setProducts] = useState<IProduct[]>([]);
+  const productIds = !value ? [] : Array.isArray(value) ? value : [value];
+  const onSelect = (product: IProduct) => {
+    if (!product) return;
+    if (mode === 'single') {
+      setProducts([product]);
+      onValueChange?.(product._id);
+      return;
+    }
+    const arrayValue = Array.isArray(value) ? value : [];
+    const isProductSelected = arrayValue.includes(product._id);
+    const newSelectedProductIds = isProductSelected
+      ? arrayValue.filter((id) => id !== product._id)
+      : [...arrayValue, product._id];
 
-  const handleSelect = (product?: IProduct) => {
-    setSelectedProduct(product);
-    onValueChange?.(product?._id || '');
-    setOpen(false);
-    scope && goBackToPreviousHotkeyScope();
+    setProducts((prevProducts) => {
+      const productMap = new Map(prevProducts.map((p) => [p._id, p]));
+      productMap.set(product._id, product);
+      return newSelectedProductIds
+        .map((id) => productMap.get(id))
+        .filter((p): p is IProduct => p !== undefined);
+    });
+    onValueChange?.(newSelectedProductIds);
   };
-
   return (
-    <PopoverScoped scope={scope} open={open} onOpenChange={setOpen} modal>
-      <Combobox.Trigger ref={ref} {...props}>
-        {!value && <Combobox.Value placeholder="Select product" />}
-        {value && <ProductInline productId={value} product={selectedProduct} />}
-      </Combobox.Trigger>
-      <Combobox.Content>
-        <SelectProductList handleSelect={handleSelect} />
-      </Combobox.Content>
-    </PopoverScoped>
+    <SelectProductContext.Provider
+      value={{
+        productIds,
+        onSelect,
+        products,
+        setProducts,
+        loading: false,
+        error: null,
+      }}
+    >
+      {children}
+    </SelectProductContext.Provider>
   );
-});
+};
 
-export const SelectProductList = ({
-  handleSelect,
-}: {
-  handleSelect: (product?: IProduct) => void;
-}) => {
+const SelectProductContent = () => {
   const [search, setSearch] = useState('');
   const [debouncedSearch] = useDebounce(search, 500);
-
-  const { loading, products, totalCount, handleFetchMore, error } = useProducts(
-    {
-      variables: {
-        searchValue: debouncedSearch,
-      },
+  const { productIds, products } = useSelectProductContext();
+  const {
+    products: productsData,
+    loading,
+    handleFetchMore,
+    totalCount,
+    error,
+  } = useProducts({
+    variables: {
+      searchValue: debouncedSearch,
     },
-  );
-
+  });
   return (
     <Command shouldFilter={false}>
-      <Command.Input value={search} onValueChange={setSearch} />
-      <Command.List>
+      <Command.Input
+        value={search}
+        onValueChange={setSearch}
+        variant="secondary"
+        wrapperClassName="flex-auto"
+        focusOnMount
+      />
+      <Command.List className="max-h-[300px] overflow-y-auto">
+        {products?.length > 0 && (
+          <>
+            {products.map((product) => (
+              <SelectProductCommandItem key={product._id} product={product} />
+            ))}
+            <Command.Separator className="my-1" />
+          </>
+        )}
         <Combobox.Empty loading={loading} error={error} />
-        {products?.map((product) => (
-          <Command.Item
-            key={product._id}
-            value={product.name}
-            onSelect={() => handleSelect(product)}
-          >
-            <ProductInline product={product} />
-          </Command.Item>
-        ))}
-        <Combobox.FetchMore
-          fetchMore={handleFetchMore}
-          currentLength={products?.length || 0}
-          totalCount={totalCount || 0}
-        />
+        {!loading &&
+          productsData
+            ?.filter(
+              (product) =>
+                !productIds.find((productId) => productId === product._id),
+            )
+            .map((product) => (
+              <SelectProductCommandItem key={product._id} product={product} />
+            ))}
+
+        {!loading && (
+          <Combobox.FetchMore
+            fetchMore={handleFetchMore}
+            currentLength={productsData?.length || 0}
+            totalCount={totalCount}
+          />
+        )}
       </Command.List>
     </Command>
   );
 };
+
+const SelectProductCommandItem = ({ product }: { product: IProduct }) => {
+  const { onSelect, productIds } = useSelectProductContext();
+  return (
+    <Command.Item
+      value={product._id}
+      onSelect={() => {
+        onSelect(product);
+      }}
+    >
+      <ProductsInline products={[product]} placeholder="Unnamed product" />
+      <Combobox.Check checked={productIds.includes(product._id)} />
+    </Command.Item>
+  );
+};
+
+const SelectProductInlineCell = ({
+  onValueChange,
+  scope,
+  ...props
+}: Omit<React.ComponentProps<typeof SelectProductProvider>, 'children'> & {
+  scope?: string;
+}) => {
+  const [open, setOpen] = useState(false);
+  return (
+    <SelectProductProvider
+      onValueChange={(value) => {
+        onValueChange?.(value);
+        setOpen(false);
+      }}
+      {...props}
+    >
+      <RecordTablePopover open={open} onOpenChange={setOpen} scope={scope}>
+        <RecordTableCellTrigger>
+          <SelectProductValue />
+        </RecordTableCellTrigger>
+        <RecordTableCellContent>
+          <SelectProductContent />
+        </RecordTableCellContent>
+      </RecordTablePopover>
+    </SelectProductProvider>
+  );
+};
+
+const SelectProductRoot = React.forwardRef<
+  React.ElementRef<typeof Combobox.Trigger>,
+  Omit<React.ComponentProps<typeof SelectProductProvider>, 'children'> &
+    React.ComponentProps<typeof Combobox.Trigger> & {
+      placeholder?: string;
+      scope?: string;
+    }
+>(({ onValueChange, className, mode, value, placeholder, scope, ...props }, ref) => {
+  const [open, setOpen] = useState(false);
+
+  return (
+    <SelectProductProvider
+    mode={mode}
+      value={value}
+
+      onValueChange={(value) => {
+        if (mode === 'single') {
+          setOpen(false);
+        }
+        onValueChange?.(value);
+      }}
+    >
+      <PopoverScoped open={open} onOpenChange={setOpen} scope={scope}>
+        <Combobox.Trigger
+          className={cn('w-full inline-flex', className)}
+          variant="outline"
+          ref={ref}
+          {...props}
+        >
+          <SelectProductValue />
+        </Combobox.Trigger>
+        <Combobox.Content>
+          <SelectProductContent />
+        </Combobox.Content>
+      </PopoverScoped>
+    </SelectProductProvider>
+  );
+})
+
+const SelectProductValue = ({placeholder}: {placeholder?: string}) => {
+  const { productIds, products, setProducts } = useSelectProductContext();
+
+  return (
+    <ProductsInline
+      productIds={productIds}
+      products={products}
+      updateProducts={setProducts}
+      placeholder={placeholder}
+    />
+  );
+};
+
+
+export const SelectProductFilterItem = () => {
+  return (
+    <Filter.Item value="product">
+      <IconShoppingCart />
+      Product
+    </Filter.Item>
+  );
+};
+
+export const SelectProductFilterView = ({
+  onValueChange,
+  queryKey,
+  mode = 'single',
+}: {
+  onValueChange?: (value: string[] | string) => void;
+  queryKey?: string;
+  mode?: 'single' | 'multiple';
+}) => {
+  const [product, setProduct] = useQueryState<string[] | string>(
+    queryKey || 'product',
+  );
+  const { resetFilterState } = useFilterContext();
+
+  return (
+    <Filter.View filterKey={queryKey || 'product'}>
+      <SelectProductProvider
+        mode={mode}
+        value={product || (mode === 'single' ? '' : [])}
+        onValueChange={(value) => {
+          setProduct(value as string[] | string);
+          resetFilterState();
+          onValueChange?.(value);
+        }}
+      >
+        <SelectProductContent />
+      </SelectProductProvider>
+    </Filter.View>
+  );
+};
+
+export const SelectProductFilterBar = ({
+  iconOnly,
+  onValueChange,
+  queryKey,
+  mode = 'single',
+}: {
+  iconOnly?: boolean;
+  onValueChange?: (value: string[] | string) => void;
+  queryKey?: string;
+  mode?: 'single' | 'multiple';
+}) => {
+  const [product, setProduct] = useQueryState<string[] | string>(
+    queryKey || 'product',
+  );
+  const [open, setOpen] = useState(false);
+
+  if (!product) {
+    return null;
+  }
+
+  return (
+    <Filter.BarItem>
+      <Filter.BarName>
+        <IconShoppingCart/>
+        {!iconOnly && 'Products'}
+      </Filter.BarName>
+      <SelectProductProvider
+        mode={mode}
+        value={product || (mode === 'single' ? '' : [])}
+        onValueChange={(value) => {
+          if (value.length > 0) {
+            setProduct(value as string[] | string);
+          } else {
+            setProduct(null);
+          }
+          setOpen(false);
+          onValueChange?.(value);
+        }}
+      >
+        <Popover open={open} onOpenChange={setOpen}>
+          <Popover.Trigger asChild>
+            <Filter.BarButton filterKey={queryKey || 'product'}>
+              <SelectProductValue />
+            </Filter.BarButton>
+          </Popover.Trigger>
+          <Combobox.Content>
+            <SelectProductContent />
+          </Combobox.Content>
+        </Popover>
+      </SelectProductProvider>
+      <Filter.BarClose filterKey={queryKey || 'product'} />
+    </Filter.BarItem>
+  );
+};
+
+export const SelectProductFormItem = ({
+  onValueChange,
+  className,
+  placeholder,
+  ...props
+}: Omit<React.ComponentProps<typeof SelectProductProvider>, 'children'> & {
+  className?: string;
+  placeholder?: string;
+}) => {
+  const [open, setOpen] = useState(false);
+  return (
+    <SelectProductProvider
+      onValueChange={(value) => {
+        onValueChange?.(value);
+        setOpen(false);
+      }}
+      {...props}
+    >
+      <Popover open={open} onOpenChange={setOpen}>
+        <Form.Control>
+          <Combobox.Trigger className={cn('w-full shadow-xs', className)}>
+            <SelectProductValue placeholder={placeholder} />
+          </Combobox.Trigger>
+        </Form.Control>
+        <Combobox.Content>
+          <SelectProductContent />
+        </Combobox.Content>
+      </Popover>
+    </SelectProductProvider>
+  );
+};
+
+
+export const SelectProduct = Object.assign(SelectProductRoot, {
+  Provider: SelectProductProvider,
+  Content: SelectProductContent,
+  Item: SelectProductCommandItem,
+  InlineCell: SelectProductInlineCell,
+  Value: SelectProductValue,
+  FilterItem: SelectProductFilterItem,
+  FilterView: SelectProductFilterView,
+  FilterBar: SelectProductFilterBar,
+  FormItem: SelectProductFormItem,
+});
