--- frontend/libs/ui-modules/src/modules/products/components/SelectProductV2.tsx
@@ -1,1 +1,0 @@
-// export const SelectProductV2 = () => {};
