+++ frontend/libs/ui-modules/src/modules/products/components/index.ts
@@ -1,2 +1,2 @@
-export * from './SelectProducts';
+export * from './SelectProduct';
 export * from './ProductsInline';
