+++ frontend/libs/ui-modules/src/modules/tags/components/SelectTags.tsx
@@ -242,13 +242,17 @@ export const TagList = ({
   );
 };
 
-export const SelectTagsValue = () => {
+export const SelectTagsValue = ({
+  placeholder,
+}: {
+  placeholder?: string;
+}) => {
   const { selectedTags, mode } = useSelectTagsContext();
 
   if (selectedTags?.length > 1) return <>{selectedTags.length} tags selected</>;
 
   return (
-    <TagList placeholder="Select tags" renderAsPlainText={mode === 'single'} />
+    <TagList placeholder={placeholder === undefined ? 'Select tags' : placeholder} renderAsPlainText={mode === 'single'} />
   );
 };
 
@@ -285,7 +289,7 @@ export const SelectTagsInlineCell = ({
     >
       <RecordTablePopover open={open} onOpenChange={setOpen} scope={scope}>
         <RecordTableCellTrigger>
-          <SelectTagsValue />
+          <SelectTagsValue placeholder=""/>
         </RecordTableCellTrigger>
         <RecordTableCellContent className="min-w-72">
           <SelectTagsContent />
